change double to decimal for csharp

+++ languages/csharp/src/main/kotlin/io/vrap/codegen/languages/csharp/CsharpBaseTypes.kt
@@ -9,7 +9,7 @@ object CsharpBaseTypes : LanguageBaseTypes(
         objectType = fromCsharpType("", "Object"),
         integerType = fromScalarCsharpType("int"),
         longType = fromScalarCsharpType("long"),
-        doubleType = fromScalarCsharpType("double"),
+        doubleType = fromScalarCsharpType("decimal"),
         stringType = fromScalarCsharpType("string"),
         booleanType = fromScalarCsharpType("bool"),
         dateTimeType = fromCsharpType("", "DateTime"),
